add storeapp, update url
