+++ finding_leftover/finding_leftover/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'post',
+    'store',
     'account.apps.AccountConfig',
 ]
