+++ finding_leftover/finding_leftover/urls.py
@@ -3,7 +3,8 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('posts/', include('post.urls')),
+    path('post/', include('post.urls')),
+    path('store/', include('store.urls')),
     path('account/', include('account.urls')),
     path('api-auth/', include('rest_framework.urls')),
 ]
