+++ finding_leftover/post/templates/store_detail.html
@@ -45,7 +45,7 @@
 </div>
 <script type="text/javascript">
 function posting(){
-    window.location.href = "{% url 'posting' %}"
+    window.location.href = "{% url 'post-create' %}"
 }
 </script>
 {% endblock %}
