+++ finding_leftover/post/urls.py
@@ -3,12 +3,10 @@ from django.urls import path, include, re_path
 from post import views
 
 urlpatterns = [
-    path('list/', views.PostListAPIView.as_view(), name="post-list"),
-    path('list/<int:pk>', views.PostDetailAPIView.as_view(), name="post-detail"),
-    # re_path('^list/search/(?P<username>.+)/$', views.SearchList.as_view()), 
-    path("storelist/", views.StoreListAPI.as_view(), name='store-list'),
-    path('store/<int:pk>', views.StoreDetailAPIView.as_view(), name="store-detail"),
-    path('posting/', views.PostCreateAPIView.as_view(), name="posting"), 
-    path('posting/<int:pk>', views.PostUpdateAPIView.as_view(), name="post-update"),
+    path('', views.PostListAPIView.as_view(), name="post-list"),
+    path('<int:pk>', views.PostDetailAPIView.as_view(), name="post-detail"),
+    
+    path('create/', views.PostCreateAPIView.as_view(), name="post-create"), 
+    path('update/<int:pk>', views.PostUpdateAPIView.as_view(), name="post-update"),
 
 ]
