+++ finding_leftover/store/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class StoreConfig(AppConfig):
+    name = 'store'
